Modification de la classe Arc

+++ src/Arc.java
@@ -1,14 +1,14 @@
 public class Arc {
     /** Attributrs */
     String nCible;
-    int poids;
+    double poids;
 
     /**
      * Constructeur
      * @param n Noeud cible de l'arc (chaine de caractere)
      * @param p Entier correspondant au poids de l'arc
      */
-    public Arc(String n, int p){
+    public Arc(String n, double p){
         this.nCible = n;
         this.poids = p;
     }
